requirements.txt
attrs==19.3.0
Automat==20.2.0
cffi==1.14.0
constantly==15.1.0
cryptography==2.9.2
cssselect==1.1.0
hyperlink==19.0.0
idna==2.10
incremental==17.5.0
itemadapter==0.1.0
lxml==4.5.2
mysqlclient==2.0.1
parsel==1.6.0
peewee==3.13.3
Protego==0.1.16
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
PyYAML==5.3.1
queuelib==1.5.0
Scrapy==2.2.1
service-identity==18.1.0
six==1.15.0
Twisted==20.3.0
w3lib==1.22.0
zope.interface==5.1.0
